line-chat/requirements.txt
fastapi
uvicorn
requests
line-bot-sdk